[test_loop] Refactor utils module in TestLoop
