+++ integration-tests/src/test_loop/tests/multinode_stateless_validators.rs
@@ -12,7 +12,8 @@ use near_primitives::views::CurrentEpochValidatorInfo;
 
 use crate::test_loop::builder::TestLoopBuilder;
 use crate::test_loop::env::TestLoopEnv;
-use crate::test_loop::utils::{execute_money_transfers, ONE_NEAR};
+use crate::test_loop::utils::transfer::execute_money_transfers;
+use crate::test_loop::utils::ONE_NEAR;
 
 const NUM_ACCOUNTS: usize = 20;
 const NUM_SHARDS: u64 = 4;
