+++ integration-tests/src/test_loop/tests/multinode_test_loop_example.rs
@@ -7,7 +7,8 @@ use near_primitives::types::AccountId;
 
 use crate::test_loop::builder::TestLoopBuilder;
 use crate::test_loop::env::TestLoopEnv;
-use crate::test_loop::utils::{execute_money_transfers, ONE_NEAR};
+use crate::test_loop::utils::transfer::execute_money_transfers;
+use crate::test_loop::utils::ONE_NEAR;
 
 const NUM_CLIENTS: usize = 4;
