+++ integration-tests/src/test_loop/utils/mod.rs
@@ -1,0 +1,4 @@
+pub mod network;
+pub mod transfer;
+
+pub(crate) const ONE_NEAR: u128 = 1_000_000_000_000_000_000_000_000;
