+++ integration-tests/src/test_loop/utils/transfer.rs
@@ -10,7 +10,7 @@ use near_primitives::transaction::SignedTransaction;
 use near_primitives::types::AccountId;
 use std::collections::HashMap;
 
-pub(crate) const ONE_NEAR: u128 = 1_000_000_000_000_000_000_000_000;
+use super::ONE_NEAR;
 
 /// Execute money transfers within given `TestLoop` between given accounts.
 /// Runs chain long enough for the transfers to be optimistically executed.
